LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollIntoView-inline-image-expected.txt

To be scrolled into view

FAIL Scrolling an inline element with a large line height uses the bounding rect assert_approx_equals: expected 594 +/- 1 but got 8
